Añadidas las vistas de favoritos y valoraciones, queda la implementación del modelo de valoración y de favoritos en el listado y el detalle de las recetas

+++ lib/src/models/valorado.dart
@@ -18,7 +18,7 @@ class Valorados {
 class Valorado {
   String email;
   int receta;
-  int valoracion;
+  double valoracion;
   String comentario;
   String fecha;
